+++ apps/scouts/src/components/boost/boostCMS/BoostCMSConfirmationPrompts/BoostCMSConfirmationPrompt.tsx
@@ -72,7 +72,7 @@ export const BoostCMSConfirmationPrompt: React.FC<{
     return (
         <section className="py-5 bg-white">
             <div className="flex flex-col items-center justify-center w-full px-4">
-                <div className="w-full flex flex-col items-center justify-center">
+                <div className="w-full flex flex-col items-center justify-center text-grayscale-900">
                     <div className="flex flex-col">
                         <div className="w-full flex items-center justify-center">
                             <h6 className="tracking-[12px] text-base font-bold text-black">